Add data-access library and base memory repository
Introduced a new data-access library to the project which includes a 'BaseMemoryRepository'. This repository abstract class together with a corresponding repository interface are designed to facilitate interaction with data storage. Additional configuration files and scripts for test, linting and building purposes have also been added as part of this update.