Refactor and improve authentication service
Removed the AppController and AppService files and added PostUser and Authentication modules to AppModule. Changed 'authentication' to 'auth' in the AuthenticationController. In the Methods of PostUserRepository and AuthenticationService made sure to return Promises and added async keyword. Also, simplified the register method in AuthenticationService, removed avatar from register and used await to fetch if a user already exists by email. Improved API endpoints support in authentication.http.