Add db:generate command and update schema output
A new command, "db:generate", has been added to the project.json file in order to generate the database using Prisma. The output path for the Prisma client has also been updated in the schema.prisma file. The new path directs the output to a dedicated folder in the node_modules directory.